KAN-94/when-opening-a-dialogue-put-selector-on-the-currently-selected-item (#66)
* refactor: add dialog selection state helpers to app

Add three helper methods to determine the current selection index for:
- Priority dialog based on active card's priority
- Sort field dialog based on current sort field setting
- Sprint dialog based on active card's sprint assignment

These methods enable dialogs to open with the selector positioned on the
currently selected item instead of always defaulting to index 0.

* refactor: create SelectionDialog trait and implementations

Add new SelectionDialog trait abstraction with three concrete implementations:

- PriorityDialog: Maps CardPriority to selection indices
- SortFieldDialog: Maps SortField to selection indices
- SprintAssignDialog: Dynamically determines sprint options and highlights
  the currently assigned sprint

This consolidates dialog rendering and selection initialization logic in one
place, replacing scattered implementations across multiple files.

* refactor: export SelectionDialog component from components module

Make the new SelectionDialog trait and implementations available to other
modules.

* refactor: use dialog selection state helpers in event handlers

Update event handlers to use the new selection state helper methods:
- get_current_priority_selection_index()
- get_current_sprint_selection_index()
- get_current_sort_field_selection_index()

This ensures dialogs open with the selector positioned on the currently
selected item across all places where dialogs are initialized.

* refactor: delegate dialog rendering to SelectionDialog components

Update three popup rendering functions to use the new SelectionDialog trait:
- render_set_card_priority_popup() → PriorityDialog::render()
- render_order_cards_popup() → SortFieldDialog::render()
- render_assign_sprint_popup() → SprintAssignDialog::render()

This centralizes dialog rendering logic and eliminates 120+ lines of duplicated
code, improving maintainability and reducing the chance of inconsistencies.

* chore: cargo fmt --all

* add changeset
