+++ .changeset/kan-94-when-opening-a-dialogue-put-selector-on-the-currently-selected-item.md
@@ -1,0 +1,10 @@
+---
+bump: patch
+---
+
+- refactor: delegate dialog rendering to SelectionDialog components
+- refactor: use dialog selection state helpers in event handlers
+- refactor: export SelectionDialog component from components module
+- refactor: create SelectionDialog trait and implementations
+- refactor: add dialog selection state helpers to app
+- Implement CardListComponent for reusable card list interactions (#65)
