+++ crates/kanban-tui/src/components/mod.rs
@@ -3,6 +3,7 @@ pub mod detail_view;
 pub mod list;
 pub mod panel;
 pub mod popup;
+pub mod selection_dialog;
 pub mod selection_list;
 
 pub use card_list_item::*;
@@ -10,4 +11,5 @@ pub use detail_view::*;
 pub use list::*;
 pub use panel::*;
 pub use popup::*;
+pub use selection_dialog::*;
 pub use selection_list::*;
