+++ crates/kanban-tui/src/ui.rs
@@ -828,31 +828,9 @@ fn render_set_card_points_popup(app: &App, frame: &mut Frame) {
 }
 
 fn render_set_card_priority_popup(app: &App, frame: &mut Frame) {
-    use kanban_domain::CardPriority;
-
-    let priorities = [
-        CardPriority::Low,
-        CardPriority::Medium,
-        CardPriority::High,
-        CardPriority::Critical,
-    ];
-
-    let selected = app.priority_selection.get();
-
-    let items: Vec<ListItem> = priorities
-        .iter()
-        .enumerate()
-        .map(|(idx, priority)| {
-            let style = if Some(idx) == selected {
-                bold_highlight()
-            } else {
-                normal_text()
-            };
-            ListItem::new(format!("{:?}", priority)).style(style)
-        })
-        .collect();
-
-    render_selection_popup_with_list_items(frame, "Set Priority", items, 30, 40);
+    use crate::components::{PriorityDialog, SelectionDialog};
+    let dialog = PriorityDialog;
+    dialog.render(app, frame);
 }
 
 fn render_card_detail_view(app: &App, frame: &mut Frame, area: Rect) {
@@ -1209,133 +1187,15 @@ fn render_set_branch_prefix_popup(app: &App, frame: &mut Frame) {
 }
 
 fn render_order_cards_popup(app: &App, frame: &mut Frame) {
-    use kanban_domain::{SortField, SortOrder};
-
-    let sort_fields = [
-        SortField::Points,
-        SortField::Priority,
-        SortField::CreatedAt,
-        SortField::UpdatedAt,
-        SortField::Status,
-        SortField::Default,
-    ];
-
-    let active_idx = sort_fields
-        .iter()
-        .position(|f| Some(*f) == app.current_sort_field);
-
-    render_selection_popup_with_lines(
-        frame,
-        "Order Tasks By",
-        Some("Select sort field:"),
-        sort_fields.iter().enumerate(),
-        |_idx, (_, field), _is_selected, is_active| {
-            let field_name = match field {
-                SortField::Priority => "Priority",
-                SortField::Points => "Points",
-                SortField::CreatedAt => "Date Created",
-                SortField::UpdatedAt => "Date Updated",
-                SortField::Default => "Task Number",
-                SortField::Status => "Status",
-            };
-
-            let order_indicator = if is_active {
-                match app.current_sort_order {
-                    Some(SortOrder::Ascending) => Some(" (↑)".to_string()),
-                    Some(SortOrder::Descending) => Some(" (↓)".to_string()),
-                    None => None,
-                }
-            } else {
-                None
-            };
-
-            (field_name.to_string(), order_indicator)
-        },
-        app.sort_field_selection.get(),
-        active_idx,
-        60,
-        50,
-    );
+    use crate::components::{SelectionDialog, SortFieldDialog};
+    let dialog = SortFieldDialog;
+    dialog.render(app, frame);
 }
 
 fn render_assign_sprint_popup(app: &App, frame: &mut Frame) {
-    let area = centered_rect(60, 50, frame.area());
-
-    frame.render_widget(Clear, area);
-
-    let block = Block::default()
-        .title("Assign to Sprint")
-        .borders(Borders::ALL)
-        .style(Style::default().bg(Color::Black));
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(2)
-        .constraints([Constraint::Length(1), Constraint::Min(0)])
-        .split(inner);
-
-    let label = Paragraph::new("Select sprint:").style(Style::default().fg(Color::Yellow));
-    frame.render_widget(label, chunks[0]);
-
-    let mut lines = vec![];
-
-    if let Some(board_idx) = app.active_board_index {
-        if let Some(board) = app.boards.get(board_idx) {
-            let board_sprints: Vec<_> = app
-                .sprints
-                .iter()
-                .filter(|s| s.board_id == board.id)
-                .collect();
-
-            let current_sprint_id = if let Some(card_idx) = app.active_card_index {
-                app.cards.get(card_idx).and_then(|c| c.sprint_id)
-            } else {
-                None
-            };
-
-            for (idx, sprint_option) in std::iter::once(None)
-                .chain(board_sprints.iter().map(|s| Some(*s)))
-                .enumerate()
-            {
-                let is_selected = app.sprint_assign_selection.get() == Some(idx);
-                let is_current = match (sprint_option, current_sprint_id) {
-                    (None, None) => true,
-                    (Some(s), Some(id)) => s.id == id,
-                    _ => false,
-                };
-
-                let style = if is_selected {
-                    Style::default().fg(Color::White).bg(Color::Blue)
-                } else if is_current {
-                    Style::default()
-                        .fg(Color::Green)
-                        .add_modifier(Modifier::BOLD)
-                } else {
-                    Style::default().fg(Color::White)
-                };
-
-                let prefix = if is_selected { "> " } else { "  " };
-                let current_indicator = if is_current { " (current)" } else { "" };
-
-                let sprint_name = if let Some(sprint) = sprint_option {
-                    sprint.formatted_name(board, board.sprint_prefix.as_deref().unwrap_or("sprint"))
-                } else {
-                    "(None)".to_string()
-                };
-
-                lines.push(Line::from(Span::styled(
-                    format!("{}{}{}", prefix, sprint_name, current_indicator),
-                    style,
-                )));
-            }
-        }
-    }
-
-    let list = Paragraph::new(lines);
-    frame.render_widget(list, chunks[1]);
+    use crate::components::{SelectionDialog, SprintAssignDialog};
+    let dialog = SprintAssignDialog;
+    dialog.render(app, frame);
 }
 
 fn render_assign_multiple_cards_popup(app: &App, frame: &mut Frame) {
